+++ paper/paper.bib
@@ -1,0 +1,106 @@
+@article{chambers2015identification,
+  title={Identification of hemagglutinin residues responsible for {H3N2} antigenic drift during the 2014--2015 influenza season},
+  author={Chambers, Benjamin S and Parkhouse, Kaela and Ross, Ted M and Alby, Kevin and Hensley, Scott E},
+  journal={Cell Reports},
+  volume={12},
+  number={1},
+  pages={1--6},
+  year={2015},
+  publisher={Elsevier}
+}
+
+@article{fowler2014deep,
+author = {Fowler, Douglas M and Fields, Stanley},
+title = {{Deep mutational scanning: a new style of protein science.}},
+journal = {Nat. Methods},
+year = {2014},
+volume = {11},
+number = {8},
+pages = {801--807},
+month = aug
+}
+
+@article{esposito2019mavedb,
+  title={MaveDB: an open-source platform to distribute and interpret data from multiplexed assays of variant effect},
+  author={Esposito, Daniel and Weile, Jochen and Shendure, Jay and Starita, Lea M and Papenfuss, Anthony T and Roth, Frederick P and Fowler, Douglas M and Rubin, Alan F},
+  journal={Genome Biology},
+  volume={20},
+  number={1},
+  pages={1--11},
+  year={2019},
+  publisher={BioMed Central}
+}
+
+@article{lee2019mapping,
+  title={Mapping person-to-person variation in viral mutations that escape polyclonal serum targeting influenza hemagglutinin},
+  author={Lee, Juhye M and Eguia, Rachel and Zost, Seth J and Choudhary, Saket and Wilson, Patrick C and Bedford, Trevor and Stevens-Ayers, Terry and Boeckh, Michael and Hurt, Aeron C and Lakdawala, Seema S and others},
+  journal={Elife},
+  volume={8},
+  year={2019},
+  publisher={eLife Sciences Publications, Ltd}
+}
+
+@article{starita2017variant,
+  title={Variant interpretation: functional assays to the rescue},
+  author={Starita, Lea M and Ahituv, Nadav and Dunham, Maitreya J and Kitzman, Jacob O and Roth, Frederick P and Seelig, Georg and Shendure, Jay and Fowler, Douglas M},
+  journal={The American Journal of Human Genetics},
+  volume={101},
+  number={3},
+  pages={315--325},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@article{wrenbeck2017deep,
+  title={Deep sequencing methods for protein engineering and design},
+  author={Wrenbeck, Emily E and Faber, Matthew S and Whitehead, Timothy A},
+  journal={Current Opinion in Structural Biology},
+  volume={45},
+  pages={36--44},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@article{gelman2019recommendations,
+  title={Recommendations for the collection and use of multiplexed functional data for clinical variant interpretation},
+  author={Gelman, Hannah and Dines, Jennifer N and Berg, Jonathan and Berger, Alice H and Brnich, Sarah and Hisama, Fuki M and James, Richard G and Rubin, Alan F and Shendure, Jay and Shirts, Brian and others},
+  journal={Genome Medicine},
+  volume={11},
+  number={1},
+  pages={85},
+  year={2019},
+  publisher={Springer}
+}
+
+@article{rubin2017statistical,
+  title={A statistical framework for analyzing deep mutational scanning data},
+  author={Rubin, Alan F and Gelman, Hannah and Lucas, Nathan and Bajjalieh, Sandra M and Papenfuss, Anthony T and Speed, Terence P and Fowler, Douglas M},
+  journal={Genome Biology},
[…]
