+++ paper/paper.md
@@ -1,0 +1,122 @@
+---
+title: '*dms-view*: Interactive visualization tool for deep mutational scanning data'
+papersize: a4
+tags:
+  - Javascript
+  - D3
+  - molecular biology
+  - protein evolution
+  - data visualization
+authors:
+  - name: Sarah K. Hilton*
+    affiliation: "1, 2"
+    orcid: 0000-0001-9278-3644
+  - name: John Huddleston*
+    affiliation: "3, 4"
+    orcid: 0000-0002-4250-2063
+  - name: Allison Black
+    affiliation: "3, 5"
+    orcid: 0000-0002-6618-4127
+  - name: Khrystyna North
+    affiliation: "1, 2"
+  - name: Adam S. Dingens
+    affiliation: 1
+    orcid: 0000-0001-9603-9409
+  - name: Trevor Bedford
+    affiliation: 3
+    orcid: 0000-0002-4039-5794
+  - name: Jesse D. Bloom
+    affiliation: "1, 2, 6"
+    orcid: 0000-0003-1267-3408
+affiliations:
+  - name: Division of Basic Sciences and Computational Biology Program, Fred Hutchinson Cancer Research Center, Seattle, WA, USA
+    index: 1
+  - name: Department of Genome Sciences, University of Washington, Seattle, WA, United States of America
+    index: 2
+  - name: Vaccine and Infectious Disease Division, Fred Hutchinson Cancer Research Center, Seattle, WA, USA
+    index: 3
+  - name: Molecular and Cell Biology, University of Washington, Seattle, WA, USA
+    index: 4
+  - name: Department of Epidemiology, University of Washington, Seattle, Washington, United States
+    index: 5
+  - name: Howard Hughes Medical Institute, Seattle, Washington, USA
+    index: 6
+date: 22 April 2020
+bibliography: paper.bib
+---
+
+# Summary and Purpose
+
+The high-throughput technique of deep mutational scanning (DMS) has recently made it possible to experimentally measure the effects of all amino-acid mutations to a protein [@fowler2014deep].
+Over the past five years, this technique has been used to study dozens of different proteins [@esposito2019mavedb] and answer a variety of research questions.
+For example, DMS has been used for protein engineering [@wrenbeck2017deep], understanding the human immune response to viruses [@lee2019mapping], and interpreting human variation in a clinical setting [@starita2017variant; @gelman2019recommendations].
+Accompanying this proliferation of DMS studies has been the development of software tools [@bloom2015software; @rubin2017statistical] and databases [@esposito2019mavedb] for data analysis and sharing.
+However, for many purposes it is important to also integrate and visualize the DMS data in the context of other information, such as the 3-D protein structure or natural sequence-variation data.
+
+Here we describe *dms-view* (https://dms-view.github.io/), a flexible, web-based, interactive visualization tool for DMS data.
+*dms-view* is written in JavaScript and [D3](https://d3js.org), and links site-level and mutation-level DMS data to a 3-D protein structure.
+The user can interactively select sites of interest to examine the DMS measurements in the context of the protein structure.
+*dms-view* tracks the input data and user selections in the URL, making it possible to save specific views of interactively generated visualizations to share with collaborators or to support a published study.
+Importantly, *dms-view* takes a flexible input data file so users can easily visualize their own DMS data in the context of protein structures of their choosing, and also incorporate additional information such amino-acid frequencies in natural alignments.
+
+Users can access *dms-view* at https://dms-view.github.io.
+The tool consists of a data section at the top and a description section at the bottom.
+The data section displays the user-specified data in three panels: the site-plot panel, the mutation-plot panel, and the protein-structure panel (\autoref{fig:fig}A).
+When sites are selected in the site-plot panel, the individual mutation values are shown in the mutation-plot panel and highlighted on the protein structure.
[…]
