### 2020-06-15 71bc091
deleted docs subdirectory
