+++ docs/github-pages/lib/github/pages.rb
@@ -1,8 +1,0 @@
-require "github/pages/version"
-
-module Github
-  module Pages
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
