+++ docs/github-pages/lib/github/pages/version.rb
@@ -1,5 +1,0 @@
-module Github
-  module Pages
-    VERSION = "0.1.0"
-  end
-end
