+++ docs/README.md
@@ -1,5 +1,0 @@
-# `dms-view` documentation
-
-This repo contains the documentation for [`dms-view`](https://jbloomlab.github.io/dms-view).
-The documentation can be found at [`dms-view.github.io/docs`](https://dms-view.github.io/docs)
-To build the site on your local computer use `bundle exec jekyll serve`.
