+++ docs/_casestudy_data/Dingens2019.md
@@ -1,18 +1,0 @@
-# [An Antigenic Atlas of HIV-1 Escape from Broadly Neutralizing Antibodies Distinguishes Functional and Structural Epitopes](https://research.fhcrc.org/content/dam/stripe/bloom/labfiles/publications/Dingens2019.pdf)
-
-Adam S. Dingens, Dana Arenz, Haidyn Weight, Julie Overbaugh, Jesse D. Bloom
-
-# Abstract 
-Anti-HIV broadly neutralizing antibodies (bnAbs)
-have revealed vaccine targets on the virus’s envelope (Env) protein and are themselves promising immunotherapies. The efficacy of bnAb-based therapies and vaccines depends in part on how readily
-the virus can escape neutralization. Although structural studies can define contacts between bnAbs
-and Env, only functional studies can define mutations
-that confer escape. Here, we mapped how all
-possible single amino acid mutations in Env affect
-neutralization of HIV by nine bnAbs targeting five epitopes. For most bnAbs, mutations at only a small
-fraction of structurally defined contact sites mediated escape, and most escape occurred at sites
-near, but not in direct contact with, the antibody.
-The Env mutations selected by two pooled bnAbs
-were similar to those expected from the combination
-of the bnAbs’s independent action. Overall, our mutation-level antigenic atlas provides a comprehensive dataset for understanding viral immune escape
-and refining therapies and vaccines.
