+++ docs/dingens2019antigenic.md
@@ -1,68 +1,0 @@
----
-layout: post
-title: Dingens et al., 2019
-permalink: /dingens2019antigenic/
----
-## [An Antigenic Atlas of HIV-1 Escape from Broadly Neutralizing Antibodies Distinguishes Functional and Structural Epitopes](https://research.fhcrc.org/content/dam/stripe/bloom/labfiles/publications/Dingens2019.pdf)
-
-Adam S. Dingens, Dana Arenz, Haidyn Weight, Julie Overbaugh, Jesse D. Bloom
-
-### Abstract
-> Anti-HIV broadly neutralizing antibodies (bnAbs)
-have revealed vaccine targets on the virus’s envelope (Env) protein and are themselves promising immunotherapies. The efficacy of bnAb-based therapies and vaccines depends in part on how readily
-the virus can escape neutralization. Although structural studies can define contacts between bnAbs
-and Env, only functional studies can define mutations
-that confer escape. Here, we mapped how all
-possible single amino acid mutations in Env affect
-neutralization of HIV by nine bnAbs targeting five epitopes. For most bnAbs, mutations at only a small
-fraction of structurally defined contact sites mediated escape, and most escape occurred at sites
-near, but not in direct contact with, the antibody.
-The Env mutations selected by two pooled bnAbs
-were similar to those expected from the combination
-of the bnAbs’s independent action. Overall, our mutation-level antigenic atlas provides a comprehensive dataset for understanding viral immune escape
-and refining therapies and vaccines.
-
-### Figure 3: Escape from V3-Glycan Supersite and V2 Apex bnAbs.
-![Figure 3](images/dingens2019antigenic_fig3.jpg)
-
->(A and B) Escape profiles for V3-glycan supersite bnAbs PGT121 (A) and 10-1074 (B). Letter heights indicate the excess fraction surviving for each mutation. Blue circles indicate structurally defined contact sites, and yellow underlines indicate a N-linked glycosylation motif. Logoplots that show escape across Env are in Data S1.  
-(C) V3-glycan supersite antibodies are shown in blue, and Env is colored according to the maximum excess fraction surviving at each site. Note that for PGT121, the closely related clonal variant PGT122 structure is used in lieu of a PGT121 structure (PDB: 5FYL and 5T3Z, respectively).   
-(D and E) Escape profiles of V2 apex antibodies PG9 (D) and PGT145 (E) presented in the same manner as (A) and (B).   
-(F) V2 apex antibodies are shown in blue, and Env is colored according to the maximum excess fraction surviving at each site (PDB: 5VJ6 and 5V8L, respectively). Median values across all biological replicates were plotted; see Figure S1 for the number of experimental replicates. See also Figures S2 and S3.
-
-Below are the `dms-view` versions:
-
[…]
