+++ docs/casestudies.md
@@ -1,14 +1,0 @@
----
-layout: post
-title: Case Studies
-permalink: /casestudies/
----
-
-# Influenza Virus hemagglutinin
-Lee _et al_., 2019 mapped the escape of IAV HA from human sera.
-Click [HERE](/docs/lee2019mapping) to explore their data using a paper figure and `dms-view`.
-
-
-# HIV Env
-Dingens _et al_., 2019 mapped the escape of HIV Env a panel of monoclonal antibodies.
-Click [HERE](/docs/dingens2019antigenic) to explore their data using paper figures and `dms-view`.

[…]
