### 2020-06-15 #2288 @skhilton
👋 @labarba: I've re-arranged our repos. Now the code and the paper are both in the repo https://github.com/dms-view/dms-view.github.io. Is there another place I should update the repo url or will this comment suffice? Thanks!
