### 2020-06-16 5436fcb
Do not interpolate data across gaps in the context panel
When data are missing for specific sites in a genome, the default area plot in
the genome context panel fills in the missing data by drawing lines between the
available data. This behavior gives the false impression that there are data
when there is actually a gap.

This commit resolves this issue by adding mock records to the long-form CSV data
for each of the gap sites and using d3 area's `defined` function [1] to only plot
area curves between data points that are defined in the original data set.

[1] https://github.com/d3/d3-shape/bl
[…]
