### 2020-06-16 22ce0d5
Collapse missing sites into those at gap boundaries
Reduces the amount of memory and time required to process mock data points for
missing data by only storing sites that occur at the beginning or end of gap
boundaries. This information is sufficient for d3's area function to plot
discontinuous areas in the context plot.
