### 2020-06-17 #2288 @csoneson
👋  @afrubin, @andrefaure - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> dms-view: Interactive visualization tool for deep mutational scanning data

Software repository: https://github.com/dms-view/dms-view.github.io
Accompanying paper: https://github.com/openjournals/joss-papers/blob/joss.02288/joss.02288/10.21105.joss.02288.pdf

JOSS reviewer guidelines and more information about the process: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
