### 2020-06-18 #2353 @whedon
[REVIEW]: dms-view: Interactive visualization tool for deep mutational scanning data
**Submitting author:** @skhilton (<a href="http://orcid.org/0000-0001-9278-3644">Sarah K. Hilton</a>)
**Repository:** <a href="https://github.com/dms-view/dms-view.github.io" target ="_blank">https://github.com/dms-view/dms-view.github.io</a>
**Version:** v1.1.0
