### 2020-06-18 #2353 @csoneson
@skhilton - I'll move my comment from the pre-review issue here so it doesn't get lost: when you get a chance, please add DOIs to the references in your paper (you can then ask whedon to generate a new article proof with `@whedon generate pdf`)
