+++ paper/paper.bib
@@ -6,18 +6,20 @@
   number={1},
   pages={1--6},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.celrep.2015.06.005},
 }
 
 @article{fowler2014deep,
 author = {Fowler, Douglas M and Fields, Stanley},
-title = {{Deep mutational scanning: a new style of protein science.}},
+title = {Deep mutational scanning: a new style of protein science.},
 journal = {Nat. Methods},
 year = {2014},
 volume = {11},
 number = {8},
 pages = {801--807},
-month = aug
+month = aug,
+doi = {10.1038/nmeth.3027},
 }
 
 @article{esposito2019mavedb,
@@ -28,7 +30,8 @@ month = aug
   number={1},
   pages={1--11},
   year={2019},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s13059-019-1845-6}
 }
 
 @article{lee2019mapping,
@@ -37,7 +40,8 @@ month = aug
   journal={Elife},
   volume={8},
   year={2019},
-  publisher={eLife Sciences Publications, Ltd}
+  publisher={eLife Sciences Publications, Ltd},
+  doi={10.7554/eLife.49324}
 }
 
 @article{starita2017variant,
@@ -48,7 +52,8 @@ month = aug
   number={3},
   pages={315--325},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ajhg.2017.07.014}
 }
 
 @article{wrenbeck2017deep,
@@ -58,7 +63,8 @@ month = aug
   volume={45},
   pages={36--44},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.sbi.2016.11.001}
 }
 
 @article{gelman2019recommendations,
@@ -69,7 +75,8 @@ month = aug
   number={1},
   pages={85},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13073-019-0698-7}
 }
 
 @article{rubin2017statistical,
@@ -80,7 +87,8 @@ month = aug
   number={1},
   pages={150},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13059-017-1272-5}
 }
 
 @Article{	  bloom2015software,
@@ -92,7 +100,8 @@ month = aug
   number	= {1},
   pages		= {1},
   year		= {2015},
-  publisher	= {BioMed Central}
+  publisher	= {BioMed Central},
+  doi = {10.1186/s12859-015-0590-4}
 }
 
 @Article{	  echave2016causes,
@@ -102,5 +111,6 @@ month = aug
 		  O},
   journal	= {Nature Reviews Genetics},
   year		= {2016},
-  publisher	= {Nature Publishing Group}
+  publisher	= {Nature Publishing Group},
+  doi = {10.1038/nrg.2015.18}
 }
