### 2020-07-02 dms-view/dms-view.github.io#153 @andrefaure
JOSS review #6
This issue is raised in context of the review process for [JOSS](https://github.com/openjournals/joss-reviews/issues/2353#issue-641122826):

Regarding functionality, dms-view (hosted at https://dms-view.github.io/) worked exactly as explained in the paper and extensive tutorial.

However, when I downloaded the code to run locally on my own machine, for some reason the mutation plot didn't render / is blank (in chrome/firefox/safari). Below I have pasted the output of the JavaScript console in chrome:

> ngl.js:1 THREE.WebGLRenderer 86
> prot_struct.js:63 Uncaught ReferenceError: createElement is not defined
>     at prot_struct.js:63
> (anonymous) @ prot_struct.js:63
> main.js:341 DOM fully loaded and parsed
> main.js:360 Waiting for promises...
> main.js:363 Promises fulfilled!
> index.html?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb:1 Access to XMLHttpRequest at 'file:///data/fonts/DejaVuSansMonoBold_SeqLogo.ttf' from origin 'null' has been blocked by CORS policy: Cross origin requests are only supported for protocol schemes: http, data, chrome, chrome-extension, https.
> main.js:351 Font could not be loaded: Font could not be loaded
> opentype.js:13946 GET file:///data/fonts/DejaVuSansMonoBold_SeqLogo.ttf net::ERR_FAILED
> pn @ opentype.min.js:1
> O.load @ opentype.min.js:1
> (anonymous) @ main.js:349
> main.js:74 Changed URL to: file:///Users/andre/Downloads/dms-view.github.io-master/index.html?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb
> main.js:128 No condition dropdown exists yet.
> main.js:197 Initialize dropdowns from URL
> main.js:55 Did not find field 'condition' in the URL.
> main.js:55 Did not find field 'site_metric' in the URL.
> main.js:55 Did not find field 'mutation_metric' in the URL.
> main.js:55 Did not find field 'selected_sites' in the URL.
> main.js:74 Changed URL to: file:///Users/andre/Downloads/dms-view.github.io-master/index.html?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb
> main.js:221 (4) ["condition", "site_metric", "mutation_metric", "selected_sites"]
> logoplot.js:60 Uncaught (in promise) TypeError: Cannot read property 'getAdvanceWidth' of undefined
>     at chart (logoplot.js:60)
>     at Pt.call (d3.v5.min.js:2)
>     at updateLogoPlot (line_plot_zoom.js:301)
>     at dropdownChange (line_plot_zoom.js:570)
>     at main.js:222
> chart @ logoplot.js:60
> call @ d3.v5.min.js:2
> updateLogoPlot @ line_plot_zoom.js:301
> dropdownChange @ line_plot_zoom.js:570
> (anonymous) @ main.js:222
> Promise.then (async)
> renderCsv @ main.js:198
> (anonymous) @ main.js:298
> Promise.then (async)
> renderDataUrl @ main.js:296
> initializeDataUrl @ main.js:329
> (anonymous) @ main.js:367
> Promise.then (async)
> (anonymous) @ main.js:361
> main.js:74 Changed URL to: file:///Users/andre/Downloads/dms-view.github.io-master/index.html?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb
