### 2020-07-02 dms-view/dms-view.github.io#154 @andrefaure
JOSS review #7
This issue is raised in context of the review process for [JOSS](https://github.com/openjournals/joss-reviews/issues/2353#issue-641122826):

I wasn't able to use links to data shared on DropBox with dms-view. When I tried this, I received the following error:

> Error reading the data URL! Please review the URL below or check the contents of the datafile by reviewing the documentation.

Perhaps further (troubleshooting) instructions could be included in the documentation / tutorials in the section "Adding your own data to the site".
