### 2020-07-02 dms-view/dms-view.github.io#155 @andrefaure
JOSS review #8
This issue is raised in context of the review process for [JOSS](https://github.com/openjournals/joss-reviews/issues/2353#issue-641122826):

A very minor point regarding the manuscript and specifically the "State of the field":

Perhaps the authors could explicitly state that there is currently no existing interactive software solving the problem they are addressing? Or perhaps the authors could mention/compare to (non-interactive) visualisation alternatives (e.g. pymol) that researchers have previously used to display residue-level scores on structures?
