### 2020-07-02 dms-view/dms-view.github.io#156 @andrefaure
JOSS review #9
This issue is raised in context of the review process for [JOSS](https://github.com/openjournals/joss-reviews/issues/2353#issue-641122826):

A very minor point regarding the documentation here on github: it wasn't immediately obvious where the "Community guidelines" for contributing were located. Perhaps the authors could include a separate section in the readme file or link therein to the corresponding markdown file ("CONTRIBUTING.md").
