### 2020-07-02 #2353 @andrefaure
Ok I have added the JOSS review issues 6-9 linked above in the target repository. In summary, dms-view is a useful and thoughtfully designed piece of software. The documentation, including case studies and tutorials, is helpful and detailed. I have completed the review checklist and the only point currently unchecked is "State of the field", although as mentioned in JOSS review issue #7 I consider this a minor point i.e. not an acceptance blocker; @csoneson shall I just go ahead and check this one too?
