### 2020-07-04 dms-view/dms-view.github.io#157 @afrubin
JOSS code cleanup
It looks like there are several duplicate files in the repo. For example, 4O5N_trimer.pdb appears in both the root and data directories.

There is also a copy of the JavaScript files in the _javascript directory, and while many of these are the same it looks like there are updated versions in the root directory.

I also noticed that the _config.yml file tests for a _data directory that doesn't exist.

However the authors choose to resolve this is fine, but including duplicate files certainly make things more confusing and harder to manage!
