### 2020-07-04 dms-view/dms-view.github.io#158 @afrubin
Suggestion for directory names
Currently the data directory includes the eraser image asset as well as the fonts. This isn't really "data" and perhaps should be moved into another directory with a suitable name.
