### 2020-07-04 dms-view/dms-view.github.io#159 @afrubin
Example PDB file
I wasn't able to find a link to the PDB record in the example documentation or the paper. After consulting the methods for [Lee et al. 2019](https://elifesciences.org/articles/49324) I see now that the "4O5N" in the file name is the PDB accession number.

This should be made more explicit in the example documentation and the authors should consider citing [Lee et al. 2014](https://www.nature.com/articles/ncomms4614) in the JOSS paper.
