### 2020-07-04 dms-view/dms-view.github.io#159 @afrubin
Looking into this further, I see that the PDB file used by dms-view is actually a trimer based on the 4O5N structure and not actually a structure from PDB directly. I was going to ask why it wasn't possible to just use a PDB accession number instead of having to provide a full file, but I see now that this is a potentially important feature.

I think it's worth adding a sentence in the paper and/or docs making this explicit (that by providing your own PDB file rather than an accession you can make these sorts of modifications).
