### 2020-07-04 dms-view/dms-view.github.io#160 @afrubin
docs repo
I'm not experienced with github-pages so maybe I'm missing something fundamental, but why are the docs in their own https://github.com/dms-view/docs repository that is separate from this one?

As someone new to the code this threw me and I think this structure causes downstream complications for versioning and doi creation.

Additionally, the current situation makes it possible (even easy) for the docs repository to contain a version of the example data that is not the same as what the tool is serving.
