### 2020-07-06 #2353 @afrubin
This is a really nice piece of software that will be very useful for the DMS community. I have raised a few issues in the repository (links below), and will probably have a bit more to say once I finish putting a MaveDB dataset into dms-view using the provided instructions.

https://github.com/dms-view/dms-view.github.io/issues/157
https://github.com/dms-view/dms-view.github.io/issues/158
https://github.com/dms-view/dms-view.github.io/issues/159
https://github.com/dms-view/dms-view.github.io/issues/160
