### 2020-07-08 cb01de9
Add support for downloading protein panel view as PNG
The protein panel's 3D graphics are rendered with Canvas and WebGL making their
export to SVG technically tricky. The library we use to render the PDB files,
NGL [1], provides a `makeImage` function that exports a PNG. We use that
function here and set the export image quality high enough that it could be used
in a lab notebook or potentially a supplemental figure.

[1] http://nglviewer.org/
