+++ logoplot.js
@@ -1,5 +1,5 @@
 function logoplotChart(selection) {
-  var colorScheme = 'functional';
+  var colorScheme = "functional";
   var divWidth = 760,
       divHeight = 250,
       margin = {top: 40, left: 40, bottom: 0, right: 0},
@@ -144,6 +144,9 @@ function logoplotChart(selection) {
       ]).nice();
 
       // Calculate the color domain.
+      if(dataToPlot[0]){
+        colorScheme =  dataToPlot[0].color_scheme
+      }
       zScale.range(d3.quantize(t => d3.interpolateSpectral(t * 0.8 + 0.1), mutations.length).reverse())
         .unknown("#cccccc");
       var functionalColors = {
@@ -170,11 +173,11 @@ function logoplotChart(selection) {
        'unknown': "#cccccc"
       };
       if(colorScheme == "functional"){
-        var colorMap = function(key){return functionalColors[key]};
         dataToPlot.forEach(function(d){
-          d["colors"] = functionalColors[d["mutation"]]
+          d["colors_mutation"] = functionalColors[d["mutation"]]
         })
       }
+      else if(colorScheme == 'custom'){}
       else{
         var colorMap = zScale;
       };
@@ -272,7 +275,7 @@ function logoplotChart(selection) {
           // desired x, y position, scaled, and then moved back by the same amount.
           return `translate(+${x} +${y}) scale(${widthScale} ${scale}) translate(-${x} -${y})`;
         })
-        .attr("fill", d => d["colors"])
+        .attr("fill", d => d["colors_mutation"])
         .on("mouseover", showTooltip)
         .on("mouseout", hideTooltip);
     });
