### 2020-07-09 db4436b
cleaned up code
