### 2020-07-09 6aef0a4
added the ability to just specify the functional color scheme
