### 2020-07-09 89bf0d1
Wait for the font to load before trying to load the rest of the app
Fixes a bug caused by my misinterpretation of opentype.js's `load` function that
I expected to return a Promise when, in fact, it has not been returning a
Promise this entire time. This commit fixes the bug by upgrading opentype.js to
the latest version where the `load` function does return a Promise as expected
and updating the downstream code to load the font from the Promise's return
values.

Related to #153
