### 2020-07-09 ee0c6b1
Serve font via https instead of http
