### 2020-07-14 3bb9616
paper edits
