+++ paper/paper.bib
@@ -114,3 +114,15 @@ doi = {10.1038/nmeth.3027},
   publisher	= {Nature Publishing Group},
   doi = {10.1038/nrg.2015.18}
 }
+
+@misc{weile,
+title={MaveVis: Visualization for MaveDB},
+journal={https://github.com/VariantEffect/mavevis},
+author={Weile, Jochen}
+}
+
+@misc{schrodinger,
+title={The PyMOL Molecular Graphics System},
+journal={https://pymol.org/2/},
+author={Schrödinger, LLC}
+}
