### 2020-07-14 dms-view/dms-view.github.io#159 @huddlej
Thank you, @afrubin! This issue has been resolved by https://github.com/dms-view/docs/pull/15.
