### 2020-07-14 939d70a
Add instructions and LaTeX template to compile paper locally
Adds a README with instructions and a template file that were in the original
manuscript repository to enable compiling to PDF locally.
