### 2020-07-14 3b84fdd
Collect fig and paper references and reorder figures
Reorders figures such that the DMS schematic is Figure 1, since it is the first
figure referenced in the manuscript.
