### 2020-07-14 1c3f689
Update references for software, rephrase statement of need
Adds references for mavevis based on the Zenodo citation [1] and for PyMOL based
on their preferred bibtex entry [2]. Also, rephrases the statement of need and
the DMS workflow caption a bit.

[1] https://zenodo.org/record/3432882/export/hx#.Xw37o1NKgWo
[2] https://pymol.org/sites/default/files/pymol.bib
