+++ paper/paper.bib
@@ -115,14 +115,26 @@ doi = {10.1038/nmeth.3027},
   doi = {10.1038/nrg.2015.18}
 }
 
-@misc{weile,
-title={MaveVis: Visualization for MaveDB},
-journal={https://github.com/VariantEffect/mavevis},
-author={Weile, Jochen}
+@software{weile,
+  author       = {Jochen Weile},
+  title        = {{VariantEffect/mavevis: v0.1.2.1: Minor release for
+                   Zenodo DOI minting}},
+  month        = sep,
+  year         = 2019,
+  publisher    = {Zenodo},
+  version      = {v0.1.2.1},
+  doi          = {10.5281/zenodo.3432882},
+  url          = {https://doi.org/10.5281/zenodo.3432882}
 }
 
-@misc{schrodinger,
-title={The PyMOL Molecular Graphics System},
-journal={https://pymol.org/2/},
-author={Schrödinger, LLC}
+@unpublished{PyMOL,
+  annote = {PyMOL
+
+The PyMOL Molecular Graphics System, Version 1.8, Schr{\"o}dinger, LLC.},
+  author = {{Schr\"odinger, LLC}},
+  date-added = {2010-08-19 17:29:55 -0400},
+  date-modified = {2015-12-22 18:04:08 -0400},
+  month = {November},
+  title = {The {PyMOL} Molecular Graphics System, Version~1.8},
+  year = {2015}
 }
