### 2020-07-14 1256e96
Link out to the contributing guide from the README
Adds a small section linking to the contributing guide and asking for any
interesting case studies folks may want to share.

Resolves #156.
