### 2020-07-14 dms-view/dms-view.github.io#153 @skhilton
@andrefaure: thanks for point this out! We've added a page to the docs describing how to run the tool locally. https://dms-view.github.io/docs/deploy-or-develop/
