### 2020-07-14 dms-view/dms-view.github.io#156 @skhilton
@andrefaure: thanks for the suggestion! We've add a link to the contributing guidelines to the README and to the docs https://dms-view.github.io/docs/deploy-or-develop/
