### 2020-07-14 b897596
Provide more information about what failed when URLs can't be loaded
Adds a place to store the server's response when the user requests a URL that
can't be loaded. This should help users debug a little more than telling them
the URL couldn't be loaded.

Related to #154.
