### 2020-07-14 dms-view/dms-view.github.io#154 @skhilton
@andrefaure: thanks for the suggestions! We've added the server error message to our `dms-view` error message when a file fails to load and we've added the reason why we can't load data from dropbox or Google Drive in the "Adding your own data to the site" section of https://dms-view.github.io/docs/tutorial/. Hopefully this helps users with debugging.
