### 2020-07-14 dfd9948
Remove old top-level data files and fix links to these from README
