### 2020-07-14 84cd013
Remove unneeded Jekyll config
We don't store data in a `_data` directory anymore, so we don't need this
config file.
