+++ _config.yml
@@ -1,2 +1,0 @@
-include:
-  - _data
