### 2020-07-14 c47ffaf
Delete old javascript directory
This directory contained duplicates of all core dms-view files that we no longer
maintain.
