+++ _javascript/logoplot.js
@@ -1,281 +1,0 @@
-function logoplotChart(selection) {
-  var colorScheme = 'functional';
-  var divWidth = 760,
-      divHeight = 250,
-      margin = {top: 40, left: 40, bottom: 0, right: 0},
-      width = divWidth - margin.left - margin.right,
-      height = divHeight - margin.top - margin.bottom,
-      fontSize = 100,
-      maxSitesToLabel = 20,
-      siteLabelInterval = 5;
-
-  // Create the base chart SVG object.
-  var svg = d3.select(selection)
-    .append("svg")
-    .attr("width", divWidth)
-    .attr("height", divHeight);
-
-  var xScale = d3.scaleBand().range(
-    [margin.left, width - margin.right]
-  ).padding(0.1);
-  var yScale = d3.scaleLinear()
-    .rangeRound([height - margin.bottom, margin.top]);
-
-  var xAxis = d3.axisBottom(xScale);
-  var yAxis = d3.axisLeft(yScale);
-
-  svg.append("g")
-    .attr("transform", `translate(0,${height - margin.bottom})`)
-    .attr("class", "x-axis")
-    .call(xAxis);
-
-  svg.append("g")
-    .attr("transform", `translate(${margin.left},0)`)
-    .attr("class", "y-axis")
-    .call(yAxis);
-
-  var zScale = d3.scaleOrdinal();
-
-  // Set x-axis label.
-  svg
-    .append("text")
-    .attr("transform", "translate(" + (divWidth / 2) + ", " + (height + 30) + ")")
-    .style("text-anchor", "middle")
-    .text("Site");
-
-  // Set y-axis label.
-  svg
-    .append("text")
-    .attr("transform", "translate(" + (12) + ", " + (height - 80) + ") rotate(-90)")
-    .style("text-anchor", "middle")
-    .attr("id", "logoplot_y_label");
-
-  // Create a genome line chart for the given selection.
-  function chart(selection) {
-    // Get the width of each letter for the requested font size. For fixed width
-    // font, the advance width should always be the same size for all
-    // letters. We calculate this width once here and reuse it during logoploto
-    // transforms later. See also, the opentype.js docs for more details:
-    // https://github.com/opentypejs/opentype.js/blob/master/README.md#fontgetadvancewidthtext-fontsize-options
-    var letterWidth = fontObject.getAdvanceWidth("X", fontSize);
-
-    selection.each(function (data) {
-      var sites = [...new Set(data.map(d => +d.site))].sort((a, b) => a - b);
-      var mutations = [...new Set(data.map(d => d.mutation))].sort();
-      if(data.length == 0){
-        var metric_name = "";
-      }
-      else {
-        var metric_name = data[0].metric_name;
-      }
-
-      xScale.domain(sites);
-      zScale.domain(mutations);
-
-      // When the width of the logo letters is less than the width of the
-      // scale's bandwidth, each letter needs to be repositioned to the right
-      // such that it centers above the corresponding axis tick. This
-      // repositioning is calculated by an x-axis offset based on half the
-      // bandwidth minus half the letter width. This offset is not needed when
-      // the bandwidth is less than the letter width, since the letters will be
-      // scaled down to the bandwidth by a transform.
-      let logoXOffset = 0;
-      if (letterWidth < xScale.bandwidth()) {
-        logoXOffset = (xScale.bandwidth() / 2) - letterWidth / 2;
-      }
-
-      // Convert long data to wide format for stacking.
-      var siteMap = new Map();
-      data.forEach(function (d) {
-        if (siteMap.get(d.site) === undefined) {
-          siteMap.set(d.site, {"site": d.site, "label": d.label_site});
-        }
-
-        siteMap.get(d.site)[d.mutation] = d.metric;
-      });
-
-      // Convert wide data map to an array.
[…]
