### 2020-07-15 b6b6fa9
Restructure repository assets into appropriate directories
Moves CSS, images, fonts, and javascript from "data" and top-level directories
into corresponding subdirectories of a new `assets/` directory and updates all
references in HTML, CSS, and javascript to these new paths.
