+++ README.md
@@ -61,8 +61,7 @@ Below are two descriptions of different combinations of dropdown menus and selec
 ### View 1 (default view)
 When [`dms-view.github.io`](https://dms-view.github.io) is loaded, the default view of the data section should look as follows:
 
-![](testing/view1.png)
-
+![](/assets/images/view1.png)
 
 [Load view 1](https://dms-view.github.io/?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&condition=2010-age-21&site_metric=site_Positive+Differential+Selection&mutation_metric=mut_Positive+Differential+Selection&selected_sites=222&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb) to explore this example.
 
@@ -72,7 +71,7 @@ The data section should look as follows when you change
 - mutation metric to "Natural Frequencies"
 - selected sites 144, 159, 192, 193, 222, 224
 
-![](testing/view2.png)
+![](/assets/images/view2.png)
 
 [Load view 2](https://dms-view.github.io/?markdown-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Flee2019mapping.md&data-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2Fflu_dms-view.csv&pdb-url=https%3A%2F%2Fdms-view.github.io%2Fdata%2FIAV%2F4O5N_trimer.pdb&selected_sites=144%2C159%2C192%2C193%2C222%2C244&condition=2009-age-65&site_metric=site_Positive+Differential+Selection&mutation_metric=mut_Natural+Frequencies) to explore this example.
 
