### 2020-07-15 39d5c87
Start migrating GitHub pages config and assets from docs repo
This allow us to test the whole site locally with Jekyll using `bundle install`
and `bundle exec jekyll serve` with docs hosted at `/new_docs/`. Fixes to broken
links, etc. to follow.
