+++ new_docs/tutorial.md
@@ -1,0 +1,145 @@
+---
+layout: post
+title: Tutorial
+permalink: /tutorial/
+---
+
+# Goal
+The goal of <a href="https://dms-view.github.io" target="_blank">`dms_view`</a> is to facilitate the analysis of deep [mutational scanning experiments](/docs/dms) through interactive visualizations.
+<a href="https://dms-view.github.io" target="_blank">`dms_view`</a> links together three pieces of information: a site-level summary metric, a mutation-level metric, and the 3-D protein structure.
+
+This tutorial consists of two sections
+1. [Introduction to the tool and its features](#introduction)
+2. [Sample analysis with Influenza Virus data](#sample-analysis)
+
+# Introduction
+
+## Layout
+
+<a href="https://dms-view.github.io" target="_blank">`dms_view`</a> is broken into two sections: the data section and the description section.
+
+The data section at the top of the page has three panels: the site panel, the mutation panel, and the protein structure panel.
+The site panel dot plot shows a summary value for each site in the experiment.
+When sites are selected, the individual mutation values are shown in the the mutation panel logoplot and the site is highlighted on the protein structure.
+You can change the data file or protein structure using the URL form fields above the panels.
+
+The description section is at the bottom of the page.
+This flexible section can be used to hold information about the experiment, give credit to those who generated the data, or hold notes about a particular analysis.
+You can change the information using the URL form field above the section.
+
+For more information on the data format for the data file, the protein structure, or the description file, please see the [data upload](/docs/dataupload) page.
+
+## How you select sites
+The goal of <a href="https://dms-view.github.io" target="_blank">`dms_view`</a> is to link site-level, mutation-level, and protein-structure information from a deep mutational scan.
+When a site is selected on the site plot, its mutation-level information is shown below as a logoplot and it is highlighted on the protein structure.
+
+### Selecting sites
+There are three ways to select a site in the site plot: clicking a site, brushing a site(s), or listing a specific site(s) by its label.
+
+1. Clicking on a site: You can click on a site using the mouse. Clicking on the site again will deselect the site.
+2. Brushing a site(s): You can brush several sites at once by clicking, holding, and swiping the mouse. Sites within the brush box will be selected.
+3. Listing a site(s): You can select a site by adding it to the selected site form field. This form field will also be updated with sites selected by clicking or brushing. The selected site field is a comma-separated list without leading or trailing spaces.
+
+### Deselecting sites
+There are four ways to deselect a site in the site plot: clicking on a site, brushing a site(s), removing a site(s) from the selected site form field, or clicking the clear selections button.
+
+1. Clicking on a site: You can click on a site using the mouse.
+2. Brushing a site(s): You can brush several sites at once to deselect them by holding the shift key and clicking, holding, and swiping the mouse. Sites within the brush box will be deselected. You can also use the radio buttons to toggle between the select and deselect brush instead of the shift-key shortcut.
[…]
