+++ new_docs/dataupload.md
@@ -1,0 +1,106 @@
+---
+layout: post
+title: How to upload your own data
+permalink: /dataupload/
+---
+
+<a href="https://dms-view.github.io" target="_blank">`dms_view`</a> takes three input files.
+1. [data file for the site plot and mutation plot](#data-file)
+2. [protein structure file ](#protein-structure)
+3. [description file](#description-file)
+
+Below are instructions on [how to construct the input files](#input-files), [how to host the input files](#hosting-input-files), and [how to load the input files to the site](#how-to-load-your-data).
+
+# Input files
+
+## Data file
+
+The data file is the main source of data for <a href="https://dms-view.github.io" target="_blank">`dms_view`</a>.
+
+This file contains the measurements for the site plot, the measurements for the mutation plot, and the map between the site plot numbering and the protein structure numbering.
+
+The data file is a `csv` (comma-separated file) and it must have the following columns:
+* **site**: The site plot will be ordered by this column. The values must be ordinal.
+* **label_site**: The site plot will be labeled by this column. The values can be numeric or a string.
+* **wildtype**: The wildtype at a given site. The tooltips will display this information. The values can be numeric or a string.
+* **mutation**: The mutation plot will display this mutation. The values must be a single uppercase letter.
+* **condition**: The experimental condition used in the experiment to generate the data for the site plot and the mutation plot. The dropdown menu will display the conditions and the user can toggle between different conditions.
+* **protein_chain**: The protein chain(s) in the pdb file for a given site. The values must be in a space-separated list. For example, populating the cell with `A C` will highlight the site on chains A and C (but not B).
+* **protein_site**: The protein site in the pdb file for a given site.
+
+The data file must contain at least one column from each of the following categories.
+
+* **site_***: The value for a given site for a given metric (represented by `*`). The values must be numeric. For example, `site_mean`.
+* **mut_***: The value for a given site for a given mutation for a given metric (represented by `*`). The values must be numeric. For example, `mut_mean`.
+
+The metric names represented by _*_ above will populate the site metric and mutation metric dropdown menus.
+
+### customizing colors in the mutation plot
+
+The default color scheme for the mutation plot indicates the biochemical property of the amino acid.
+If you would like to use custom colors in the mutation plot, you should include a **color_for_mutation** column.
+The cells in the **color_for_mutation** column should contain either 1) a valid hex color code or 2) the keyword "functional" to indicate the default color scheme.
+This allows you mix-and-match which site/mutation/site metric combinations use a custom color and which use the default color scheme.
+Invalid hex color codes or empty cells will default to black.
+
+## example: HIV antibody escape
+
+Imagine an experiment to test the effect of alanine mutations on the ability of antibodies to bind to HIV.
+In this small experiment, the effect of an alanine mutation was measured at position 1 or position 2 for antibody A and antibody B.
+
+Here is what the data file might look like for this experiment.
+
+site|site_label|wildtype|mutation|condition|protein_chain|protein_site|mut_max|mut_mean|site_median|
+---|---|---|---|---|---|---|---|---|---|
+1|pos1|G|A|Ab1|A|27|5|3.5|7
+1|pos1|G|A|Ab2|A|27|0|0|7
+2|pos2|D|A|Ab1|A|27|0|0|45
+2|pos2|D|A|Ab2|A|27|100|50|45
+
+Note how the the **site_*** values are repeated for each mutation at that site.
+For a real world example, see [the default data file used by `dms_view`](https://raw.githubusercontent.com/dms-view/dms-view.github.io/master/data/IAV/flu_dms-view.csv).
+
+## protein structure
+
+The protein structure file is a [protein data bank (pdb) file](https://en.wikipedia.org/wiki/Protein_Data_Bank_(file_format)).
[…]
