+++ new_docs/index.md
@@ -1,0 +1,17 @@
+---
+layout: default
+---
+
+# Welcome to `dms-view`
+
+The goal of `dms-view` is to facilitate easy exploration of deep mutational scanning experiments through interactive visualizations.
+
+![Quick walkthrough of dms-view](frontpage.gif)
+
+## Getting started
+
+If you want to learn how to use the tool or explore pre-existing datasets, check out the [tutorial](./tutorial) or the [case studies](./casestudies).
+
+If you would like to upload your own data, check out [upload your own data](dataupload) or [deep mutational scanning primer](dms).
+
+If you would like to learn more about the tool, checkout the [preprint](https://www.biorxiv.org/content/10.1101/2020.05.14.096842v1) describing `dms-view`.
