### 2020-07-16 #2353 @csoneson
👋 Just checking in again to see where we're at.
@thorvere - could you give an update on the status of your review (I see some opened issues in the software repository, but no checked boxes in the checklist above)?
@skhilton - it looks like you've gotten some good feedback (and already worked through some of it). Let us know if you have questions, and when you feel like you're ready for the reviewers to have a second look.
Thanks everyone!
