### 2020-07-16 69a94fd
Update layouts, remove permalinks, and update images references
Updates HTML and Markdown such that all references and links resolve as expected
inside the temporary `new_docs` directory. The most important change here is the
removal of "permalinks" and updates to image paths.

Minor edits to link text are also included here.
