### 2020-07-16 4a21626
removed references to github.com/dms-view/docs
