### 2020-07-17 db99899
handle null values for protein site
