### 2020-07-17 cb6fd32
handle missing chains as well
