### 2020-07-17 ed98009
Check for non-missing protein chain instead of attribute of chain
Corrects a logical test that could never evaluate to true and prevented any
sites from appearing selected on the protein view.
