+++ assets/js/line_plot_zoom.js
@@ -275,7 +275,7 @@ function genomeLineChart() {
 
     // update the PROTEIN structure
     if(!missingData.includes(circleData.protein_site) &&
-       !missingData.includes(circleData.protein_chain.protein_site)){
+       !missingData.includes(circleData.protein_chain)){
       circleData.protein_chain.forEach(function(chain){
         selectSiteOnProtein(":" + chain + " and " +
           circleData.protein_site,
