### 2020-07-17 17fdf15
handle blank chains
