### 2020-07-21 aa4a215
different color scheme on representation change
