### 2020-07-21 9d8e279
correct color scheme on load
