### 2020-07-21 c104263
make repeated code into a function
