### 2020-07-21 811c53c
moved color definitions to main.js
