### 2020-07-22 803d2ed
fixed minor bug in the color scheme
