+++ assets/js/main.js
@@ -19,7 +19,7 @@ var clearbuttonchange;
 let protein;
 const greyColor = "#999999";
 const targetChainsColor = greyColor;
-const alternativeChainsColor = 'black';
+const alternativeChainsColor = '#555555';
 
 var fontPath = "/assets/fonts/DejaVuSansMonoBold_SeqLogo.ttf";
 var fontObject;
