### 2020-07-25 b2db645
added checkbox to toggle dropdown menu
