### 2020-07-28 41fff41
changed dropdown to data fields
