+++ CHANGELOG.md
@@ -1,0 +1,31 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/):
+
+Types of changes
+- `Added` for new features.
+- `Changed` for changes in existing functionality.
+- `Deprecated` for soon-to-be removed features.
+- `Removed` for now removed features.
+- `Fixed` for any bug fixes.
+- `Security` in case of vulnerabilities.
+
+## [Unreleased]
+
+### Added
+- Download images of each panel
+- Added custom color scheme option for mutation panel
+
+### Changed
+- Reorganized assets
+- Moved `docs` directory into main repo
+
+### Fixed
+- Fixed bug when protein chain or site was a missing value
+
+## [1.0.0] - 2020-05-15
+
+### Added
+- Initial submission to JOSS
