### 2020-07-28 2772f79
added error message if invalid color
