### 2020-07-29 c7253b3
Rename protein color fields to reflect their context
