+++ assets/js/main.js
@@ -11,7 +11,7 @@ let csvDataUrl;
 let conditiondropdown;
 let sitedropdown;
 let mutdropdown;
-const dropdownsToTrack = ["condition", "site_metric", "mutation_metric", "selected_sites"];
+const dropdownsToTrack = ["condition", "site_metric", "mutation_metric", "selected_sites", "protein-data-color", "protein-other-color"];
 
 var dropdownChange;
 var clearbuttonchange;
@@ -288,12 +288,19 @@ function renderPdb(data, dataUrl) {
   protein = data;
   protein.setRotation([2, 0, 0])
   protein.autoView()
-  colorWholeProtein(protein, polymerSelect.value, false)
 
   // If data have been loaded into the site plot, select any sites from that
   // panel in the protein view, too.
-  if (chart !== undefined) {
+  if (chart.data !== undefined) {
     chart.updateSites(d3.selectAll(".selected").nodes().map(d => d3.select(d)));
+
+    // If the CSV data are already loaded, we can try to color chains.
+    colorWholeProtein(protein, polymerSelect.value, true);
+  }
+  else {
+    // If we haven't loaded any CSV data yet, we don't know which chains will be
+    // colored, so we use the default.
+    colorWholeProtein(protein, polymerSelect.value, false);
   }
 
   return protein;
