### 2020-07-29 24d583e
error message works for either color
