### 2020-07-29 1f49072
added custom protein colors to the docs
