### 2020-08-04 f555614
filter out empty string chain name
