### 2020-08-10 #2353 @thorvere
dms-view is a well-designed visualization tool allowing to interactively analyze DMS data in the context of protein structures.  Upon addressing some minor modifications for which I opened issues, I think it is a very useful contribution for the scientific community. Thus, I can recommend it for publication.
