### 2020-08-10 #2353 @csoneson
@openjournals/dev - any idea what could cause this compilation issue? 🤔
