### 2020-08-10 #2353 @csoneson
@skhilton - could you check this?

> It seems there's a problem rendering the dms.pdf file.
https://github.com/dms-view/dms-view.github.io/blob/master/paper/dms.pdf
Maybe it's not a well formatted pdf?
