### 2020-08-10 #2353 @skhilton
@csoneson: This should be fixed now! One of the pdfs was not formatted correctly. The version on the master branch complies nicely using https://whedon.theoj.org/ for me. Let me know if there is trouble on your end
