### 2020-08-11 60309f5
fixed broken link in the custom colors section
