### 2020-08-12 dms-view/dms-view.github.io#179 @afrubin
Bad links on main page
The links for "using the tool" and "uploading your own data" on the main page are returning 404 errors. They lead to https://dms-view.github.io/docs/tutorial/ and https://dms-view.github.io/docs/dataupload/ respectively. It seems like for whatever reason, the trailing `/` is causing the problem.
