### 2020-08-12 #2353 @afrubin
@csoneson Looks great to me!

@skhilton I did find https://github.com/dms-view/dms-view.github.io/issues/179 when was checking over the changes, but that should be easy to handle.

The manuscript cites MaveVis using its Zenodo DOI, but as it was initially described in Esposito et al. (the MaveDB paper) you could use that citation instead if you prefer.
