### 2020-08-12 0cff157
Fix broken links from dms-view main page
Resolves #179.
