### 2020-08-12 705308d
Cite Esposito et al. for MaveVis in addition to the Zenodo DOI
It can be helpful to have the citation for the original paper that software was
developed for and the software DOI itself, so we include both here.
