### 2020-08-13 #2353 @csoneson
👋 @skhilton - I submitted a pull request with a couple of minor fixes to the text. Please check and merge if it looks ok. For consistency, could you represent the country in all affiliations in the same way (now there's USA, United States and United States of America 😃)? Also, it would be good to indicate somewhere what the asterisk means.

The next steps are the following:
- [ ] Create a tagged release of the software, and report back the version tag here
- [ ] Archive the release in Zenodo
- [ ] Check that the Zenodo archive has the same title and author list as the paper.
- [ ] Report back the Zenodo DOI here
