### 2020-08-13 #2353 @huddlej
@csoneson, @skhilton and I intended for the asterisks to indicate shared co-first authorship. Is there a more formal way to represent this authorship status for JOSS, or is it sufficient to add a legend below the authorship section describing what the asterisk means? Thank you!
