### 2020-08-14 58cdb80
Standardize state and country in paper affiliations
Use "WA" for "Washington" and "USA" for the "United States of America".
