### 2020-08-14 7da3bbe
Indicate shared first authorship with a footnote
Uses JOSS's recommended approach for indicating co-authorship [1] by adding an
inline footnote to Sarah's name and a superscript asterisk to John's name to
recreate the same style of asterisk that appears in the footnote. This
superscript asterisk is a hack to get the same footnote appearance, since using
a named footnote (as supported by pandoc [2]) causes the footnote's contents to
be duplicated and use different markers.

[1] https://github.com/openjournals/joss-reviews/issues/2353#issuecomment-674027760
[2] https://pandoc.org/MANUAL.html#foo
[…]
