### 2020-08-14 e59d78b
Update CHANGELOG.md
