### 2020-08-14 #2353 @csoneson
@huddlej - the way we often recommend for indicating shared co-first authorship is to include a footnote, e.g., as done here: https://joss.theoj.org/papers/10.21105/joss.02145
