### 2020-08-14 #2353 @skhilton
@csoneson: The new tagged version of the software is v1.1.0 (https://github.com/dms-view/dms-view.github.io/releases/tag/v1.1.0) and the zenodo doi is 10.5281/zenodo.3986155. @huddlej updated the affiliations. Let us know if there are any issues!
