### 2020-08-17 #2353 @huddlej
@csoneson Thank you for all of your help with this review process! I just noticed a minor issue in the author list on the right-hand navigation of the paper's page (https://joss.theoj.org/papers/10.21105/joss.02353) where my name is listed as "John Huddleston$". I think this has to do with the Latex code I added to place an asterisk next to my name.

Is it possible to change this to remove the "$" from the end somehow? Should I open a separate issue for this? Thanks again!
