### 2020-08-17 #2353 @kyleniemeyer
@openjournals/dev can you help @huddlej update the paper?
