### 2020-08-17 #2353 @skhilton
@csoneson, @kyleniemeyer: Could we also change the link software repository link from https://github.com/jbloomlab/dms-view-ms to https://github.com/dms-view/dms-view.github.io/. During the review process we moved the manuscript to the main software repository. Thanks!
